api路径修改；签和接口修改

+++ business/business-order/src/main/java/com/funtl/myshop/plus/business/dto/SignOffParamDto.java
@@ -17,7 +17,7 @@ public class SignOffParamDto implements Serializable {
     private Long ordersFDetailAuto = 0L;
 
     @ApiModelProperty(value = "试算单号")
-    private Long ordersAuto;
+    private Integer ordersAuto;
 
     @ApiModelProperty(value = "签核意见")
     private String memo;
