+++ provider/provider-order-service/src/main/java/com/funtl/myshop/plus/provider/mapper/WorkFlowDocMapper.java
@@ -8,7 +8,7 @@ import tk.mybatis.mapper.MyMapper;
 import java.util.List;
 
 public interface WorkFlowDocMapper extends MyMapper<WorkFlowDoc> {
-    SignOffList selectSignOffList(@Param("docPostID") Long docPostID, @Param("roleId") Integer roleId);
+    SignOffList selectSignOffList(@Param("docPostID") Integer docPostID, @Param("roleId") Integer roleId);
 
     List<WorkFlowDoc> selectWorkFlowDoc(@Param("docPostID") Long docPostID, @Param("roleId") Integer roleId, @Param("type") Integer type);
 }
